Catch all modern AVRs, not just tinyAVR and improve timer selection
The existing tests were not reliably identifying when megaTinyCore was in use, and did nothing to handle other modern AVRs that should be treated in the same way (DxCore for Dx-series and megaAVR 0-series). Now these are detected and use the same platform. I suggest renaming it more appropriately, but that's a decision that shouldn't be up to me.
Relating to these parts
* Arduino.h is now included. There are cases where it is required for code compatibility between families owing to Microchip's willingness to change how they spell bitfield names..
* The timer selection logic previously ensured that tone could never be used with RadioHead. Now try to avoid that. If we have TCB1 and it's not used for millis, use that, otherwise, see if we have TCB3 (ie, 48+pin DA/DB-series) and use that  if we do (since TCB2 is the default, and the user went out of their way to move millis off of TCB2, so they probably don't want this library to use it either) Only if neither of those are an option do we take TCB0.